upates to pipelins and widget file loading
updates pipelinse to now be stored in a Surface. This will likely change in the futer but for the time being is needed for things.

Updates widget loading to be able to grab all toml files from ether the defaulted widgets folder in the root virt folder or from a provided path in the new config.toml file in the root virt folder.